napi fogyasztasi adatok int-rol float-ra a pontosabb megjelenites erdekeben

+++ src/main/java/overheadcost/overheadcost/entities/GasChartModel.java
@@ -9,5 +9,5 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class GasChartModel {
     private String date;
-    private int consumption;
+    private float consumption;
 }
